Use intptr_t instead of NSInteger.

+++ src/classes/GRMustacheContext.m
@@ -727,7 +727,7 @@ NSString *canonicalKeyForKey(Class klass, NSString *key);
     });
     
     Class klass = object_getClass(object);
-    NSInteger result = (int)CFDictionaryGetValue(cache, klass);   // 0 = undefined, 1 = YES, 2 = NO
+    intptr_t result = (intptr_t)CFDictionaryGetValue(cache, klass);   // 0 = undefined, 1 = YES, 2 = NO
     if (!result) {
         result = class_conformsToProtocol(klass, @protocol(GRMustacheTagDelegate)) ? 1 : 2;
         CFDictionarySetValue(cache, klass, (const void *)result);
